+++ moldyn/AddClusterColours.cpp
@@ -193,7 +193,7 @@ bool moldyn::AddClusterColours::getDataCallback(Call& caller) {
 //                            (parts.GetMaxColourIndexValue() - parts.GetMinColourIndexValue());
                         unsigned int clusterId = (unsigned int)(values[j]/100000000.0);
                         unsigned int colorIdx = (clusterId%13);
-                        float v = colorIdx/13.0;
+                        float v = static_cast<float>(colorIdx) / 13.0f;
                         if (v < 0.0f) v = 0.0f;
                         else if (v > 1.0f) v = 1.0f;
